# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: iRespond API
  version: 0.2.0
  description: >-
    Mobile-first community action API. Reported needs are observations until
    verification. YugabyteDB is the relational system of record and RustFS is
    the canonical S3 object store for evidence bytes.
servers:
  - url: https://api.irespond.example
...
paths:
...
  /v1/needs/{id}/project:
    post:
      tags: [Projects]
      operationId: convertNeedToProject
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/NeedId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConvertProjectRequest' }
      responses:
        '201': { description: Action project created, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Projects]
      operationId: getProjectByNeed
      parameters: [{ $ref: '#/components/parameters/NeedId' }]
      responses:
        '200': { description: Action project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/projects/{id}:
    get:
      tags: [Projects]
      operationId: getProject
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      responses:
        '200': { description: Project detail, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/projects/{id}/contribution-needs:
    post:
      tags: [Projects, Contributions]
      operationId: addContributionNeed
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContributionNeedRequest' }
      responses:
        '201': { description: Contribution need created, content: { application/json: { schema: { $ref: '#/components/schemas/ContributionNeed' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/projects/{id}/contribution-needs/{needId}/offers:
    post:
      tags: [Contributions]
      operationId: offerContribution
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - { $ref: '#/components/parameters/ContributionNeedId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContributionOfferRequest' }
      responses:
        '201': { description: Offer created, content: { application/json: { schema: { $ref: '#/components/schemas/ContributionOffer' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/projects/{id}/milestones:
    post:
      tags: [Projects]
      operationId: addProjectMilestone
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MilestoneRequest' }
      responses:
        '201': { description: Milestone created, content: { application/json: { schema: { $ref: '#/components/schemas/Milestone' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/projects/{id}/milestones/{milestoneId}/transition:
    post:
      tags: [Projects]
      operationId: transitionProjectMilestone
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - { $ref: '#/components/parameters/MilestoneId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StateRequest' }
      responses:
        '200': { description: Milestone transitioned, content: { application/json: { schema: { $ref: '#/components/schemas/Milestone' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/projects/{id}/roles/invite:
    post:
      tags: [Projects]
      operationId: inviteProjectRole
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleInviteRequest' }
      responses:
        '201': { description: Role invitation created, content: { application/json: { schema: { $ref: '#/components/schemas/RoleInvite' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/project-role-invites/{inviteId}/accept:
    post:
      tags: [Projects]
      operationId: acceptProjectRoleInvite
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/InviteId' }]
      responses:
        '200': { description: Role accepted, content: { application/json: { schema: { $ref: '#/components/schemas/RoleInvite' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/projects/{id}/transition:
    post:
      tags: [Projects]
      operationId: transitionProject
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StateRequest' }
      responses:
        '200': { description: Project transitioned, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: StratoID/OIDC RS256 access token.
    internalBearer:
      type: http
      scheme: bearer
      description: Dedicated internal notification ingestion token.
  parameters:
    NeedId: { name: id, in: path, required: true, schema: { type: string } }
    ProjectId: { name: id, in: path, required: true, schema: { type: string } }
...
    ContributionNeedId: { name: needId, in: path, required: true, schema: { type: string } }
...
    MilestoneId: { name: milestoneId, in: path, required: true, schema: { type: string } }
    InviteId: { name: inviteId, in: path, required: true, schema: { type: string } }
...
  responses:
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Authentication required or invalid
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Authenticated identity lacks required authority
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: State or integrity conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ConvertProjectRequest:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        ownerCommunityId: { type: string }
    Project:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        needId: { type: string }
        title: { type: string }
        state: { type: string }
    ContributionNeedRequest:
      type: object
      required: [kind, description]
      properties:
        kind: { type: string }
        description: { type: string }
        quantityNote: { type: string }
    ContributionNeed:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        projectId: { type: string }
        kind: { type: string }
    ContributionOfferRequest:
      type: object
      properties:
        note: { type: string }
        availabilityNote: { type: string }
    ContributionOffer:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        projectId: { type: string }
        contributorId: { type: string }
        state: { type: string }
...
    MilestoneRequest:
      type: object
      required: [title, sequence]
      properties:
        title: { type: string }
        description: { type: string }
        sequence: { type: integer }
        targetAt: { type: [string, 'null'], format: date-time }
    Milestone:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        projectId: { type: string }
        state: { type: string }
    StateRequest:
      type: object
      required: [state]
      properties: { state: { type: string } }
    RoleInviteRequest:
      type: object
      required: [actorId, role]
      properties:
        actorId: { type: string }
        role: { type: string }
    RoleInvite:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        projectId: { type: string }
        actorId: { type: string }
        role: { type: string }
...